[Refactor] #20 UserController, UserService 리팩토링
